Успешно вырезали плюсовую часть хэндлера
